Updated Room.java, breaking the changes from last commit into more methods

+++ src/Animate.java
@@ -5,7 +5,7 @@ public interface Animate
     public boolean performAction(int i);
 
     public boolean receiveDamage(int damage, String type);
-    public int getHealth();
+    public float getHealth();
     public Inventory getInventory();
     public int getAttackDamage();
     public int getWisdom();
